requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
consonance==0.1.5
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
cssselect2==0.7.0
cups==0.0.6
daphne==4.1.2
Django==5.0.7
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-timezone-field==7.0
fonttools==4.54.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
mercadopago==2.2.3
msgpack==1.1.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.28.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pyphen==0.16.0
python-axolotl==0.2.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
zope.interface==7.0.3
zopfli==0.2.3
